Store error message in exception object
This was necessary, because git_lasterror was not always set for all errors.
In cases where lasterror is not set, we need to call git_strerror() instead,
and we need to store the result somewhere. Alternatively, we could store the
error code, and call strerror() in "what", but this is a better approach.