Fix buggy behavior in QGitOId
length() returns length as number of hex chars (because this is what
the lookup_prefix functions expect.

Internal data structure correctly resized to half the number of
hex chars, because that is the size of the binary representation.

Array resized after calling the parse function, because it actually
writes to all 20 bytes, zeroing what remains after parsing the given
chars.

The truncation resulting in dividing an odd number of chars by two
correctly corresponds to the last char being ignored by libgit2 if
length is odd.