Repositories can be cloned using SSH public key authentication
A new class - Credentials - wraps the git_cred concept of libgit2. The
design is a bit different from libgit2: where libgit2 uses a callback
function and the user can do there anything she wants the Credentials
class offers only the most usual cases. Currently it only implements the
SSH public key authentication type but other types are easy to add.

Created a dedicated user account to github for the clone test. The
public and private keys for that account are included in this commit.
Those keys shouldn't off course be used for anything else than testing.