Added git push
Pushing to a remote is now possible. A Push object is created for a
remote using the Repository class. The Push object is then configured
as needed. Finally calling the execute() method does the actual pushing.

The test needed a bit of configuration since libgit2 supports pushing to
only bare local repositories - a bare local repository is created before
the pushing test is run.