Fixed some includes
Let the qgit2.h file be the only special header in the root directory.
It's a header that shouldn't be used within the library itself, it's
just a shorthand for the users of the library.

All other headers (and sources) are under the src directory.

The root directory isn't added as an include directory during building
anymore.