Added diffing between two trees in a repository
A new method in Repository - diffTrees() - can be used to produce an
object of a new class: Diff. The diffTrees() method is a direct wrapper
to libgit2's git_diff_tree_to_tree() method and the Diff class is a
wrapper to the git_diff type.

The Diff object can be used to get the deltas from the diff. The needed
classes - DiffDelta and DiffFile - were already present in libqgit2
beforehand.