libqgit2 needs initialization before use
Introduced an initialization and a shutdown method for libqgit2. They
are basically just wrappers for the libgit2's corresponding methods. The
only added feature is that the libgit2 initialization method doesn't get
called more than once. There's a bug in libgit2 0.21.0 that would result
into resource leaks otherwise.

The initialization is required for HTTPS transport. That's why the clone
test calls the initialization method. Basically it should be called in
every test...