Added checkout HEAD function to Repository
Added proper checkout options too. The only place where this is now used
is the new checkoutHead() function but it could (and probably should) be
used in other places too where checkouting happens.

The CheckoutOptions class only reveals a limited subset of the options
in the git_checkout_options struct of libgit2.