Changed Repository to use private implementation
The bigger change in this commit is that the git_repository pointer is
always (when necessary) accessed in a safe way. In every place where a
valid pointer is needed (libgit2 asserts in many functions if it doesn't
get a valid pointer) an exception is thrown if no valid pointer is
available.