Fix build for Qt4
The "signals" macro is defined as "protected" in Qt4 which makes it not
OK to emit signals of a class from outside the class (as was done in the
now removed RepositoryCloneRemoteListener). But the Repository Private
class has access to it's owner class and can emit its signals too. So
the Private class can act as the RemoteListener too (and now there's no
need to create a temporary object for the duration of the clone either).