Added a helper class for converting between different path encodings.
On Windows, the path names that come out of libgit2 are all encoded as
UTF-8, and the path names given as input are also expected to be encoded
as UTF-8. This is due to the POSIX emulating functions in libgit2 using
UTF-8 for everything. However, since QFile's encodeName/decodeName
convert between UTF-8 and local encoding, which is not UTF-8, I've used
QString's toUtf8/fromUtf8 methods instead.

In Linux and OS X the POSIX functions most probably use the system
locale so it is probably best to use QFile's methods in systems other
than Windows.