Avoid setting AUTOMOC property on libgit2
Setting CMAKE_AUTOMOC before adding libgit2 subdirectory sets the
AUTOMOC property on libgit2. This seems to be confusing at least
to MSVC, and results in a compilation error. We handle this by
setting the AUTOMOC property on libqgit2 only. Also, AUTOMOC works
for Qt 4 as well as Qt 5, so there is no need to handle things
differently for Qt 4.